refactor: move signature verification functions to EpochManagerAdapter  (#7656)
Part of https://github.com/near/nearcore/issues/6910, depends on #7655

Must be reviewed per commit.
