+++ chain/epoch-manager/src/adapter.rs
@@ -1,6 +1,10 @@
 use near_chain_primitives::Error;
+use near_crypto::Signature;
 use near_primitives::{
+    block_header::{Approval, ApprovalInner, BlockHeader},
+    errors::EpochError,
     hash::CryptoHash,
+    sharding::{ChunkHash, ShardChunkHeader},
     types::{
         validator_stake::ValidatorStake, AccountId, ApprovalStake, BlockHeight, EpochId, ShardId,
     },
@@ -64,6 +68,81 @@ pub trait EpochManagerAdapter: Send + Sync {
         last_known_block_hash: &CryptoHash,
         account_id: &AccountId,
     ) -> Result<(ValidatorStake, bool), Error>;
+
+    fn verify_block_vrf(
+        &self,
+        epoch_id: &EpochId,
+        block_height: BlockHeight,
+        prev_random_value: &CryptoHash,
+        vrf_value: &near_crypto::vrf::Value,
+        vrf_proof: &near_crypto::vrf::Proof,
+    ) -> Result<(), Error>;
+
+    /// Verify validator signature for the given epoch.
+    /// Note: doesnt't account for slashed accounts within given epoch. USE WITH CAUTION.
+    fn verify_validator_signature(
+        &self,
+        epoch_id: &EpochId,
+        last_known_block_hash: &CryptoHash,
+        account_id: &AccountId,
+        data: &[u8],
+        signature: &Signature,
+    ) -> Result<bool, Error>;
+
+    /// Verify signature for validator or fisherman. Used for validating challenges.
+    fn verify_validator_or_fisherman_signature(
+        &self,
+        epoch_id: &EpochId,
+        last_known_block_hash: &CryptoHash,
+        account_id: &AccountId,
+        data: &[u8],
+        signature: &Signature,
+    ) -> Result<bool, Error>;
+
+    /// Verify header signature.
+    fn verify_header_signature(&self, header: &BlockHeader) -> Result<bool, Error>;
+
+    /// Verify chunk header signature.
+    /// return false if the header signature does not match the key for the assigned chunk producer
+    /// for this chunk, or if the chunk producer has been slashed
+    /// return `Error::NotAValidator` if cannot find chunk producer info for this chunk
+    /// `header`: chunk header
+    /// `epoch_id`: epoch_id that the chunk header belongs to
+    /// `last_known_hash`: used to determine the list of chunk producers that are slashed
+    fn verify_chunk_header_signature(
+        &self,
+        header: &ShardChunkHeader,
+        epoch_id: &EpochId,
+        last_known_hash: &CryptoHash,
+    ) -> Result<bool, Error> {
+        self.verify_chunk_signature_with_header_parts(
+            &header.chunk_hash(),
+            header.signature(),
+            epoch_id,
+            last_known_hash,
+            header.height_created(),
+            header.shard_id(),
+        )
+    }
+
+    fn verify_chunk_signature_with_header_parts(
+        &self,
+        chunk_hash: &ChunkHash,
+        signature: &Signature,
+        epoch_id: &EpochId,
+        last_known_hash: &CryptoHash,
+        height_created: BlockHeight,
+        shard_id: ShardId,
+    ) -> Result<bool, Error>;
+
+    /// Verify aggregated bls signature
+    fn verify_approval(
+        &self,
+        prev_block_hash: &CryptoHash,
+        prev_block_height: BlockHeight,
+        block_height: BlockHeight,
+        approvals: &[Option<Signature>],
+    ) -> Result<bool, Error>;
 }
 
 /// A technical plumbing trait to conveniently implement [`EpochManagerAdapter`]
@@ -154,5 +233,138 @@ impl<T: HasEpochMangerHandle + Send + Sync> EpochManagerAdapter for T {
         let fisherman = epoch_manager.get_fisherman_by_account_id(epoch_id, account_id)?;
         let block_info = epoch_manager.get_block_info(last_known_block_hash)?;
         Ok((fisherman, block_info.slashed().contains_key(account_id)))
+    }
+
+    fn verify_block_vrf(
+        &self,
+        epoch_id: &EpochId,
+        block_height: BlockHeight,
+        prev_random_value: &CryptoHash,
+        vrf_value: &near_crypto::vrf::Value,
+        vrf_proof: &near_crypto::vrf::Proof,
+    ) -> Result<(), Error> {
+        let epoch_manager = self.read();
+        let validator = epoch_manager.get_block_producer_info(epoch_id, block_height)?;
+        let public_key = near_crypto::key_conversion::convert_public_key(
+            validator.public_key().unwrap_as_ed25519(),
+        )
+        .unwrap();
+
+        if !public_key.is_vrf_valid(&prev_random_value.as_ref(), vrf_value, vrf_proof) {
+            return Err(Error::InvalidRandomnessBeaconOutput);
+        }
+        Ok(())
+    }
+
+    fn verify_validator_signature(
+        &self,
+        epoch_id: &EpochId,
+        last_known_block_hash: &CryptoHash,
+        account_id: &AccountId,
+        data: &[u8],
+        signature: &Signature,
+    ) -> Result<bool, Error> {
+        let (validator, is_slashed) =
+            self.get_validator_by_account_id(epoch_id, last_known_block_hash, account_id)?;
+        if is_slashed {
+            return Ok(false);
+        }
+        Ok(signature.verify(data, validator.public_key()))
+    }
+
+    fn verify_validator_or_fisherman_signature(
+        &self,
+        epoch_id: &EpochId,
+        last_known_block_hash: &CryptoHash,
+        account_id: &AccountId,
+        data: &[u8],
+        signature: &Signature,
+    ) -> Result<bool, Error> {
+        match self.verify_validator_signature(
+            epoch_id,
+            last_known_block_hash,
+            account_id,
+            data,
+            signature,
+        ) {
+            Err(Error::NotAValidator) => {
+                let (fisherman, is_slashed) =
+                    self.get_fisherman_by_account_id(epoch_id, last_known_block_hash, account_id)?;
+                if is_slashed {
+                    return Ok(false);
+                }
+                Ok(signature.verify(data, fisherman.public_key()))
+            }
+            other => other,
+        }
+    }
+
+    fn verify_header_signature(&self, header: &BlockHeader) -> Result<bool, Error> {
+        let epoch_manager = self.read();
+        let block_producer =
+            epoch_manager.get_block_producer_info(header.epoch_id(), header.height())?;
+        match epoch_manager.get_block_info(header.prev_hash()) {
+            Ok(block_info) => {
+                if block_info.slashed().contains_key(block_producer.account_id()) {
+                    return Ok(false);
+                }
+                Ok(header.signature().verify(header.hash().as_ref(), block_producer.public_key()))
+            }
+            Err(_) => return Err(EpochError::MissingBlock(*header.prev_hash()).into()),
+        }
+    }
+
+    fn verify_chunk_signature_with_header_parts(
+        &self,
+        chunk_hash: &ChunkHash,
+        signature: &Signature,
+        epoch_id: &EpochId,
+        last_known_hash: &CryptoHash,
+        height_created: BlockHeight,
+        shard_id: ShardId,
+    ) -> Result<bool, Error> {
+        let epoch_manager = self.read();
+        let chunk_producer =
+            epoch_manager.get_chunk_producer_info(epoch_id, height_created, shard_id)?;
+        let block_info = epoch_manager.get_block_info(last_known_hash)?;
+        if block_info.slashed().contains_key(chunk_producer.account_id()) {
+            return Ok(false);
+        }
+        Ok(signature.verify(chunk_hash.as_ref(), chunk_producer.public_key()))
+    }
+
+    fn verify_approval(
+        &self,
+        prev_block_hash: &CryptoHash,
+        prev_block_height: BlockHeight,
+        block_height: BlockHeight,
+        approvals: &[Option<Signature>],
+    ) -> Result<bool, Error> {
+        let info = {
+            let epoch_manager = self.read();
+            epoch_manager.get_all_block_approvers_ordered(prev_block_hash).map_err(Error::from)?
+        };
+        if approvals.len() > info.len() {
+            return Ok(false);
+        }
+
+        let message_to_sign = Approval::get_data_for_sig(
+            &if prev_block_height + 1 == block_height {
+                ApprovalInner::Endorsement(*prev_block_hash)
+            } else {
+                ApprovalInner::Skip(prev_block_height)
+            },
+            block_height,
+        );
+
+        for ((validator, is_slashed), may_be_signature) in info.into_iter().zip(approvals.iter()) {
+            if let Some(signature) = may_be_signature {
+                if is_slashed || !signature.verify(message_to_sign.as_ref(), &validator.public_key)
+                {
+                    return Ok(false);
+                }
+            }
+        }
+        Ok(true)
     }
 }
